Sales Cycle Updated

+++ engineering/engineering/doc_events/purchase_order.py
@@ -146,7 +146,8 @@ def cancel_sales_order(self):
 	if self.so_ref:
 		so = frappe.get_doc("Sales Order", self.so_ref)
 		so.flags.ignore_permissions = True
-		so.cancel()
+		if so.docstatus == 1:
+			so.cancel()
 
 		url = get_url_to_form("Sales Order", so.name)
 		frappe.msgprint(_("Sales Order <b><a href='{url}'>{name}</a></b> has been cancelled!".format(url=url, name=so.name)), title="Sales Order Cancelled", indicator="red")
@@ -158,5 +159,6 @@ def delete_sales_order(self):
 
 		frappe.db.set_value("Sales Order", self.so_ref, 'po_ref', '')
 
-		frappe.delete_doc("Sales Order", self.so_ref, force = 1, ignore_permissions=True)
-		frappe.msgprint(_("Sales Order <b>{name}</b> has been deleted!".format(name=self.so_ref)), title="Sales Order Deleted", indicator="red")
+		if frappe.db.exists("Sales Order", self.so_ref):
+			frappe.delete_doc("Sales Order", self.so_ref, force = 1, ignore_permissions=True)
+			frappe.msgprint(_("Sales Order <b>{name}</b> has been deleted!".format(name=self.so_ref)), title="Sales Order Deleted", indicator="red")
